+++ src/lib/Scenes/City/Components/AllEvents.tsx
@@ -3,6 +3,7 @@ import { EventSection } from "lib/Scenes/City/Components/EventSection"
 import { BucketKey, BucketResults } from "lib/Scenes/Map/Bucket"
 import React from "react"
 import { FlatList, ViewProperties } from "react-native"
+import { BMWEventSection } from "./BMWEventSection"
 import { FairEventSection } from "./FairEventSection"
 import { SavedEventSection } from "./SavedEventSection"
 
@@ -65,6 +66,13 @@ export class AllEvents extends React.Component<Props, State> {
       })
     }
 
+    if (buckets.museums && buckets.museums.length) {
+      sections.push({
+        type: "bmw",
+        data: buckets.museums,
+      })
+    }
+
     if (buckets.closing && buckets.closing.length) {
       sections.push({
         type: "closing",
@@ -95,6 +103,7 @@ export class AllEvents extends React.Component<Props, State> {
   }
 
   renderItem = ({ item: { data, type } }) => {
+    const { cityName } = this.props
     switch (type) {
       case "fairs":
         return <FairEventSection data={data} />
@@ -106,6 +115,8 @@ export class AllEvents extends React.Component<Props, State> {
         return <EventSection title="Opening shows" data={data} />
       case "closing":
         return <EventSection title="Closing shows" data={data} />
+      case "bmw":
+        return <BMWEventSection title="BMW Art Guide" data={data} cityName={cityName} />
       case "saved":
         return <SavedEventSection data={data} />
       case "header":